requirements.txt
fastapi==0.103.1
uvicorn[standard]==0.23.2
python-multipart==0.0.6
google-cloud-texttospeech==2.14.1
google-cloud-speech==2.19.0
google-cloud-aiplatform==1.27.1
google-generativeai==0.3.1
PyPDF2==3.0.1
python-docx==0.8.11
pandas==2.0.1
redis==4.5.5
pydantic==2.0.3
python-jose==3.3.0
